fix: refactor code based on code reviews
- Move code to compareNames function
- Add additional check for see if display name and legal name is equal
