+++ lib/collective.lib.js
@@ -181,10 +181,43 @@ export const canContributeRecurring = (collective, user) => {
  * Displays the name string as "Legal name (Display name)" if legal name exists.
  * Example: Sudharaka (Suds)
  */
-export const concatenateDisplayNameWithLegalName = (legalName, displayName) => {
-  if (legalName) {
+export const formatAccountName = (legalName, displayName) => {
+  if (legalName && legalName !== displayName) {
     return `${legalName} (${displayName})`;
   } else {
     return displayName;
   }
+};
+
+/*
+ * Validate the account holder name against the legal name. Following cases are considered a match,
+ *
+ * 1) Punctuation are ignored; "Evil Corp, Inc" and "Evil Corp, Inc." are considered a match.
+ * 2) Accents are ignored; "François" and "Francois" are considered a match.
+ * 3) The first name and last name order is ignored; "Benjamin Piouffle" and "Piouffle Benjamin" is considered a match.
+ * 4) If one of account holder name or legal name is not defined then this function returns true.
+ */
+export const compareNames = (accountHolderName, legalName) => {
+  if (legalName && accountHolderName) {
+    const namesArray = legalName.split(' ');
+    let legalNameReversed;
+    if (namesArray.length === 2) {
+      const firstName = namesArray[0];
+      const lastName = namesArray[1];
+      legalNameReversed = `${lastName} ${firstName}`;
+    }
+    if (
+      accountHolderName.localeCompare(legalName, undefined, {
+        sensitivity: 'base',
+        ignorePunctuation: true,
+      }) &&
+      accountHolderName.localeCompare(legalNameReversed, undefined, {
+        sensitivity: 'base',
+        ignorePunctuation: true,
+      })
+    ) {
+      return false;
+    }
+  }
+  return true;
 };
